[MODIFIED] `removeWallet(0)` removes the wallet with index 0, not all wallets

+++ src/services/wallet.js
@@ -267,7 +267,7 @@ export function getHDAccounts() {
  * deleteAccount(); // Delete all accounts
  */
 export function deleteAccount(accPointer) {
-    if(!accPointer) {
+    if(!accPointer && accPointer !== 0) {
         deleteByIndex(null);
         console.log(allAccounts);
     }
